requirements.txt
aioboto3==12.3.0
aiobotocore==2.11.2
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.34
botocore==1.34.34
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cssselect==1.2.0
cycler==0.12.1
deprecation==2.1.0
entrypoints==0.4
extra-streamlit-components==0.1.70
Faker==23.2.1
fastapi==0.110.0
favicon==0.7.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.49.0
frozenlist==1.4.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
htbuilder==0.6.2
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
prometheus_client==0.20.0
protobuf==4.25.3
psycopg2==2.9.9
pyarrow==15.0.0
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pymdown-extensions==10.7
pyparsing==3.1.1
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
s3transfer==0.10.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.27
st-annotated-text==4.0.1
starlette==0.36.3
streamlit==1.31.1
streamlit-authenticator==0.3.1
streamlit-calendar==1.1.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.3
streamlit-modal==0.1.2
streamlit-quill==0.0.3
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.27.1
validators==0.22.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
